Updated cmdImg_node.cpp
Fixed all of the minor bugs that came with the last patch; however, with the addition of proper image taking into the node, its processing speed has reduced tremendously. I am unsure of why this is, but I believe it has something to do with the axis_camera and its constituents. Will have to check again on a later date.